+++ src/AdapterGoogleCloud.ts
@@ -83,6 +83,13 @@ export class AdapterGoogleCloud extends AbstractAdapter {
       if (options.public === true) {
         await this._client.bucket(name, options).makePublic();
       }
+      if (options.versioning === true) {
+        await this._client.bucket(name).setMetadata({
+          versioning: {
+            enabled: true,
+          },
+        });
+      }
       return { value: "ok", error: null };
     } catch (e) {
       return { value: null, error: e.message };
@@ -140,7 +147,7 @@ export class AdapterGoogleCloud extends AbstractAdapter {
     }
     const expires = exp.valueOf();
     try {
-      const file = this._client.bucket(bucketName).file(fileName);
+      const file = this._client.bucket(bucketName).file(decodeURI(fileName));
       const url = (await file.getSignedUrl({
         action: "read",
         expires,